pyproject.toml
[tool.poetry]
name = "nr-app"
version = "0.1.0"
description = ""
authors = ["Dmitry Figol <git@dmfigol.me>"]

[tool.poetry.dependencies]
python = "^3.9"
nornir = "^3"
colorama = "*"
textfsm = "^1"
matplotlib = "^3"
requests = "^2"
networkx = "^2"
lxml = "^4"
jupyter = "^1"
httpx = "*"
nornir-utils = "*"
nornir-jinja2 = "*"
scrapli = {extras = ["ssh2", "textfsm", "paramiko", "genie", "asyncssh"], version = "*", allow-prereleases = true }
scrapli-netconf =  { version = "*", allow-prereleases = true }
scrapli-cfg =  { git = "https://github.com/scrapli/scrapli_cfg.git", branch = "main", allow-prereleases = true }
nornir-scrapli = { git = "https://github.com/scrapli/nornir_scrapli", branch = "develop"}
"ruamel.yaml" = "*"
genie = "^21.3"
pyats = "^21.3"

[tool.poetry.dev-dependencies]
bpython = "*"
pdbpp = "*"
black = {version = "*", allow-prereleases = true}
flake8 = "*"
mypy = "*"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
